Error handing done

+++ backend/routes/post-routes.js
@@ -24,7 +24,7 @@ const storage=multer.diskStorage({
     cb(error, "backend/images")
   },
   filename:(req, file, cb)=> {
-    console.log("Hello world");
+    //console.log("Hello world");
     const name=file.originalname.toLowerCase().split(' ').join('-');
     const ext=MIME_TYPE_MAP[file.mimetype];
     cb(null, name+'-'+Date.now()+'.'+ext)
@@ -32,10 +32,13 @@ const storage=multer.diskStorage({
 })
 router.post("",AuthCheck, multer({storage:storage}).single("image"),(req,res,next)=>{
   const url=req.protocol+"://"+req.get("host");
+
+ console.log(req.userData);
   const post=new Post({
          title:req.body.title,
          content:req.body.content,
-         imagePath:url+"/images/"+req.file.filename
+         imagePath:url+"/images/"+req.file.filename,
+         creator:req.userData.userId
   });
   post.save().then(createdPost=>{
     console.log(createdPost)
@@ -48,6 +51,10 @@ router.post("",AuthCheck, multer({storage:storage}).single("image"),(req,res,nex
      }
 
     });
+  }).catch(err=>{
+    res.status(500).json({
+      message:"Post Creation Failed",
+    })
   });
  console.log(post);
 })
@@ -70,7 +77,11 @@ router.get("",(req,res,next)=>{
     posts:fetchPosts,
     maxPosts:count
  })
- })
+ }).catch(err=>{
+  res.status(500).json({
+    message:"Fetching Post Failed",
+  })
+})
 })
 router.put('/:id',AuthCheck ,multer({storage:storage}).single("image"),(req,res,next)=>{
   const url=req.protocol+"://"+req.get("host");
@@ -78,18 +89,30 @@ router.put('/:id',AuthCheck ,multer({storage:storage}).single("image"),(req,res,
          _id:req.params.id,
          title:req.body.title,
          content:req.body.content,
-         imagePath:url+"/images/"+req.file.filename
+         imagePath:url+"/images/"+req.file.filename,
+
   });
  const post=new Post({
    _id:req.params.id,
    title:req.body.title,
    content:req.body.content,
-   imagePath:postData.imagePath
+   imagePath:postData.imagePath,
+   creator:req.userData.userId
  })
- Post.updateOne({_id:req.params.id},post).then(response=>{
+ Post.updateOne({_id:req.params.id,creator:req.userData.userId},post).then(response=>{
    console.log(response)
-   res.status(200).json({message:"Post Updated Successfully"});
- })
+   if(response.nModified>0){
+    res.status(200).json({message:"Post Updated Successfully"});
+   }
+   else{
+   res.status(401).json({message:"Not Authorized"})
+   }
+
+ }).catch(err=>{
+  res.status(500).json({
+    message:"Post Updating Failed",
+  })
+})
 })
 router.get('/:id',(req,res,next)=>{
   Post.findById(req.params.id).then(post=>{
@@ -101,6 +124,10 @@ router.get('/:id',(req,res,next)=>{
     // console.log("post fetched not found")
       res.status(400).json({message:"Post not Found"})
     }
+  }).catch(err=>{
+    res.status(500).json({
+      message:"Fetching Post Failed",
+    })
   })
 
 })
@@ -108,10 +135,21 @@ router.get('/:id',(req,res,next)=>{
 router.delete("/:id",AuthCheck,(req,res,next)=>{
  const id=req.params.id;
  console.log(id);
- Post.deleteOne({_id:id}).then(result=>{
+ Post.deleteOne({_id:id,creator:req.userData.userId}).then(result=>{
    console.log(result);
-   res.status(200).json({message:"Post deleted Successfully"})
- })
+
+   if(result.n>0){
+    res.status(200).json({message:"Post deleted Successfully"})
+   }
+ else{
+  res.status(401).json({message:"Not Authorized"})
+ }
+
+ }).catch(err=>{
+  res.status(500).json({
+    message:"Post Deletion Failed",
+  })
+})
 // res.json("Post deleted from backend")
 })
